# src/main/resources/api/service_openapi.yaml
openapi: 3.0.3
info:
  description: API definition for the CDA
  version: 3.0.0
  title: CDA API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

# testing only
#servers:
#  - url: https://cancerdata.dsde-dev.broadinstitute.org
#  - url: https://cda.cda-dev.broadinstitute.org
#  - url: http://localhost:8080
#  - url: http://35.192.60.10:8080
...
paths:
  /status:
    get:
      summary: CDA status
      description: Returns the operational status of the service
      operationId: serviceStatus
      tags:
        - meta
      responses:
        200:
          $ref: "#/components/responses/StatusResponse"
        500:
          $ref: "#/components/responses/StatusResponse"
...
components:
...
  schemas:
...
    SystemStatus:
      type: object
      properties:
        ok:
          type: boolean
          description: status of this service
        systems:
          type: object
          additionalProperties:
            type: object
            properties:
              ok:
                type: boolean
              messages:
                type: array
                items:
                  type: string
...
  responses:
    StatusResponse:
      description: common status response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SystemStatus"
...